Move task API to a separate file
